refactor: link from ensnode app to ensnode-common package
