+++ apps/ensnode/package.json
@@ -15,6 +15,7 @@
   "packageManager": "pnpm@9.12.0",
   "dependencies": {
     "@ensdomains/ensjs": "^4.0.2",
+    "ensnode-common": "workspace:*",
     "ensnode-subgraph-graphql": "workspace:*",
     "hono": "^4.6.14",
     "ponder": "^0.8.26",
