+++ apps/ensnode/src/api/index.ts
@@ -1,6 +1,6 @@
 import { ponder } from "ponder:registry";
 import { default as schema } from "ponder:schema";
-import { graphql as subgraphGraphQL } from "ensnode-subgraph-graphql";
+import { graphql as subgraphGraphQL } from "ensnode-subgraph-graphql/middleware";
 import { graphql as ponderGraphQL } from "ponder";
 
 // use ponder middleware at root
