+++ apps/ensnode/src/plugins/linea.eth/handlers/EthRegistrar.ts
@@ -1,10 +1,7 @@
 import { ponder } from "ponder:registry";
 import schema from "ponder:schema";
+import { makeSubnodeNamehash, tokenIdToLabel } from "ensnode-common/subname-helpers";
 import { zeroAddress } from "viem";
-import {
-  makeSubnodeNamehash,
-  tokenIdToLabel,
-} from "../../../../../../packages/ensnode-core/src/subname-helpers";
 import { makeRegistrarHandlers } from "../../../handlers/Registrar";
 import { upsertAccount } from "../../../lib/db-helpers";
 import { ownedName, pluginNamespace } from "../ponder.config";
