+++ packages/ensnode-common/src/helpers.ts
