--- packages/ensnode-core/tsconfig.json
@@ -1,26 +1,0 @@
-{
-  "compilerOptions": {
-    // Type checking
-    "strict": true,
-    "noUncheckedIndexedAccess": true,
-
-    // Interop constraints
-    "verbatimModuleSyntax": false,
-    "esModuleInterop": true,
-    "isolatedModules": true,
-    "allowSyntheticDefaultImports": true,
-    "resolveJsonModule": true,
-
-    // Language and environment
-    "moduleResolution": "bundler",
-    "module": "ESNext",
-    "noEmit": true,
-    "lib": ["ES2022"],
-    "target": "ES2022",
-
-    // Skip type checking for node modules
-    "skipLibCheck": true
-  },
-  "include": ["./**/*.ts"],
-  "exclude": ["node_modules"]
-}
