+++ packages/ensnode-subgraph-graphql/src/index.ts
@@ -1,0 +1,1 @@
+export * from "./middleware";
